+++ src/main/java/dao/OrderDao.java
@@ -79,5 +79,19 @@ public class OrderDao {
 	}
 	
 	// pointhistory & user 포인트 수정하는 sql문
+	
+	// 주문 여부를 확인하는 sql
+	public int getOrderCount(int productNo,int userNo) throws SQLException {
+		String sql = "select count(*) cnt "
+				+ "from sul_orders O, sul_order_items I "
+				+ "where O.order_no = I.order_no "
+				+ "and I.pd_no = ? "
+				+ "and O.user_no = ? ";
+		
+		return helper.selectOne(sql, rs -> {
+			return rs.getInt("cnt");
+			
+		},productNo,userNo);
+	}
 
 }
