+++ src/main/java/dao/ProductReviewDao.java
@@ -53,19 +53,6 @@ public class ProductReviewDao {
 		},productNo);
 	}
 	
-	public int getReviewCount(int productNo, int userNo) throws SQLException {
-		String sql = "select count(*) cnt "
-				+ "from sul_orders O, sul_order_items I "
-				+ "where O.order_no = I.order_no "
-				+ "and I.pd_no = ? "
-				+ "and O.user_no = ? ";
-		
-		return helper.selectOne(sql, rs -> {
-			return rs.getInt("cnt");
-			
-		},productNo,userNo);
-	}
-	
 	public int getReviewUserCount(int productNo) throws SQLException {
 		String sql = "select count(*) cnt "
 				+ "from sul_reviews R, sul_users U "
