+++ src/main/webapp/product/detail.jsp
@@ -154,12 +154,42 @@
 				<div class="row border-bottom mb-3">
 	   				<div class="col-2 p-3 ">
 	         			<p class="text-muted mb-1"><%=review.getCreatedDate() %></p>
+	         			<p class="text-muted mb-1">평점 : 
+					    <%
+							if(review.getScore() == 5){
+						%>
+							<span>★★★★★</span>
+						<%
+							} else if(review.getScore() ==4 ) {
+						%>
+							<span>★★★★</span>
+						<%
+							} else if(review.getScore() == 3) {
+						%>
+							<span>★★★</span>
+						<%
+							} else if (review.getScore() == 2) {
+						%>
+							<span>★★</span>
+						<%
+							} else if (review.getScore() ==1 ) {
+						%>
+							<span>★</span>
+						<%
+							}
+						%>
+						</p>
 	        	 		<p class="text-muted"><%=review.getUserId() %></p>
 					</div>
-	      			<div class="col-10 p-3">
-	         			<p class="small"><%=review.getContent() %> </p>
-	         			<p><a href="">1</a>개의 댓글이 있습니다. <span class="text-info">추천 </span> : <span class="test-info"><%=review.getLikeCount() %></span> <a href="" class="btn btn-info btn-sm">추천하기</a></p>
-	     		 	</div>
+		      		<div class="col-10 p-3">
+		         			<p class="small"><%=review.getContent() %> </p>
+		         		<div>
+							<img alt="" src="../images/sample1.jpg" class="img-thumbnail" width="100">
+						</div>
+		         			<p><a href="">1</a>개의 댓글이 있습니다. <span class="text-info ">추천 </span> : <span class="test-info"><%=review.getLikeCount() %></span>
+		         				<%--비 로그인시 리뷰 추천창이 안뜨도록 했습니다. 리뷰 추천기능 넣었습니다. --%>
+		         			   <a href="" class="btn btn-info btn-sm <%=user == null ? "btn-outline-secondary disabled" : "btn-outline-primary" %>" onclick="likeReview(<%=review.getNo() %>)">추천하기</a></p>
+		     		</div>
 	   			</div>
 			<%
 					}
@@ -279,6 +309,34 @@
 
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/js/bootstrap.bundle.min.js"></script>
 <script type="text/javascript">
+	
+	// 추천체크 기능입니다.
+	function likeReview(reviewNo) {
+		
+		let xhr = new XMLHttpRequest();
+		xhr.onreadystatechange = function() {
+			if (xhr.readyState === 4 && xhr.status === 200) {
+				let jsonText = xhr.responseText;
+				let result = JSON.parse(jsonText);
+	
+				if(result.same) {
+					alert("자신이 작성한 글은 추천할수 없습니다.");
+					return;
+				} 
+				if(result.already) {
+					alert("이미 추천한 리뷰입니다");
+					return;
+				}
+				location.href="pdreviewLike.jsp?reviewNo="+ reviewNo;
+				alert("추천이 완료되었습니다");
+				return;
+				
+			}
+	
+		}
+		xhr.open("GET", '../review/reviewLikeCheck.jsp?reviewNo=' + reviewNo);
+		xhr.send();
+	}
 	// 상품문의 글쓰기 체크와 모두 완료시 속성값 주는 기능
 	function questionCheckForm() {
 		// form의 id를 통해 값을 가져옵니다.
