--- src/main/webapp/product/productQuestionlist.jsp
@@ -1,20 +1,0 @@
-<%@page import="dto.QuestionDto"%>
-<%@page import="java.util.List"%>
-<%@page import="dao.ProductQuestionDao"%>
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-</head>
-<body>
-<%
-	ProductQuestionDao productQuestionDao = ProductQuestionDao.getInstance();
-	
-%>
-
-	</div>
-</body>
-</html>
