+++ src/main/webapp/product/questionlist.jsp
@@ -1,0 +1,18 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>상품문의 전체보기</title>
+<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet">
+</head>
+<body>
+<jsp:include page="../common/nav.jsp">
+	<jsp:param name="menu" value="qlist"/>
+</jsp:include>
+	<%
+		
+	%>
+</body>
+</html>
